requirements.txt
asgiref==3.5.2
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
future==0.18.2
importlib-metadata==4.12.0
Markdown==3.4.1
psycopg2==2.9.3
pytz==2022.2.1
serializers==0.2.4
sqlparse==0.4.2
tzdata==2022.1
zipp==3.8.1
